sdk: move key origin implementation to file
* inline HDKey initialization in tests
