+++ crates/sdk/src/signer/mod.rs
@@ -2,8 +2,11 @@
 pub mod core;
 /// Signer-specific error enumerations capturing execution constraints and mapping internal failure types.
 pub mod error;
+/// Contains abstractions that allow the `Signer` to remain agnostic to the origin of its keys.
+mod key_origin;
 /// Utilities for injecting witness data bindings into Simplicity environments.
 mod wtns_injector;
 
-pub use core::{HDKey, KeyOrigin, Signer, SignerTrait, SingleKey};
+pub use core::{Signer, SignerTrait};
 pub use error::SignerError;
+pub use key_origin::{HDKey, KeyOrigin, SingleKey};
